requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==5.3.0

# Django
django==2.2.9

# Django rest-framework
djangorestframework==3.9.0 

# Sql
sqlparse==0.3.0

#mysqlclient
mysqlclient==2.0.1

#django-import-export
django-import-export==2.4.0

# dotenv
python-dotenv

# webpack
django-webpack-loader==0.7.0


